more box-undo debugging
Not really sure why this is acting odd. I'm tired and not thinking very
clearly.

Definitely a good bit of cleanup where I'm standardizing on a way to
record position and revert moves for any "mover" (using a stored px/py
and x/y).

But depending on when I'm recording the position of a box, I'm lagging
behind (no good), or missing something so undos can make the player and
box be on the same space (easy to do if you push a box against a wall,
then move away from the box, then undo).

Just odd.