turn position into a whole object
Right now it's a little silly because it's just x/y, and then I had to
make some extra functions for table comparison and duplication. But in
the end it does simplify things to just compare the pos and prevpos
tables, or assign one to be a dupe of the other. Or compare two pos
tables.

And then later it will help if position has more than x and y, like what
sub-level you're on?

Also remove some extra record_pos(box) calls, which I was threw in when
I was confused about how/where I was recording the position of boxes
badly for undo.