some graphics tweaks
Boxes are just plain rectangles for now - made rectdim util function
take a parameter to decided whether or not to fill

And make the background sprite a smaller rectangle, so you can see the
separation between it and a target rectangle if that target is the same
color. And then the target gets highlighted if it has a box.