start getting "sublevels" going
This is the main thing that sets parabox apart and what I'm trying to
figure out — entering a sublevel from within a level, and doing it
freely (entering, exiting, pushing things in and out).

It's a bit of a brain twister right now, but at least I've gotten this
far with allowing a level to have any idea of levels, and that a box
represents one. And now where you would enter it!