getting closer w/ sublevels
Drawing the pixelated-sublevel-in-box vertically centered feels better.

Removing the map-slice and camera bits so I can see everything that's
happening. That's pretty important.

Adjusting the position of boxes/targets/player based on what level
they're in, so they appear in the right place. I'll see if that
continues to be correct when I bring back the map-slice and camera
stuff.

Also, need to make this consistent. It looks like sometimes the player
will enter the sublevel, and sometimes they won't. And it's not like
sometimes when you go on the box, but sometimes based on rerunning the
game. No idea on that right now.

This has a bunch of debugging code that's either left in, or commented
out. Lots of cleanup to follow, once this is definitely working.