some more sublevel work
Removing any idea of being "out of bounds", because that's actually
going to be how to exit from a sublevel. And putting a little space
between the level and sublevel, so it's possible to exit that sublevel
and wander around. (That space might have to stay, or it may be
something I can remove if I get the exiting working.)

Doing some extraction/refactoring of checking for a position being
blocked, and checking it again when entering a sublevel, so it's not
possible to enter from the wrong place.

Also supposedly correctly adjusting screen position from map position
and level. I'm sure this will never be a problem again.