add level shown in parabox trailer
This is a little wacky, but it seems to work. The big thing is making it
so a box can be pushed into a sublevel. And it wasn't too hard? Kind of?

It's not completely like the trailer, because I don't have a player
target. But that should be simple enough to add.